score board and keyboard takes score now

+++ Assets/Scripts/ScoreBoard/ScoreboardEntryData.cs
@@ -6,6 +6,6 @@ namespace DapperDino.Scoreboards
     public struct ScoreboardEntryData
     {
         public string entryName;
-        public int entryScore;
+        public float entryScore;
     }
 }
